Simplify AST nodes by using more enums
All AST nodes are now either an enum or a tuple-like struct, and unlike
before all enums now directly encode their information in a tuple where
required (instead of referencing a struct). This reduced the amount of
new types defined in the AST source file.

These changes were also mirrored in the to_verilog module, so that the
single test still passes.

The to_verilog function no longer takes a codegen struct, which
simplifies all the calls to it while we still don't need to reference
additional tables / data. This may have to change in the future, but
also may be (mostly) replaced by the syntax visitor that will be
implemented.