Change integration test and add shift operator support
The and_gate integration test (which just tested that a minimal Silica
module could compiile at all) was replaced with an expressions test,
which should test the possibly types of expressions supported by Silica
in a single module.

Support was added for the shift operator, making sure that (just like in
Verilog) it has precedence over the bitwise operators.

Fixed an issue where parentesized expressions weren't put in parenthesis
in the output Verilog code, which could change their meaning.