Add errors and testing to lexer
A lexer error type was added to the lexer module, and implementations
were further fleshed out. Lexer now implements Iterator in a way that
should be friendly to the lalrpop parser.

Unit testing was added to the lexer module, which already caught a bug
where numbers were not allowed in an identifier token.