Simplify AST by removing nodes for Identifier and Literal
These nodes are now just a variation of an Expression, and all the
necessary changes have been made to all other files such that all tests
pass.

This will simplify situations where it is necessary to create an AST or
AST node by hand, such as when writing unit tests, by reducing the
amount of indirection occuring.