Change syntax to a subcrate and update testing
Move the syntax module into its own subcrate. The module is large
enough where it may benefit from being its own subcrate for compile
times, and will also help with modularization of the codebase.

Add an error type to the syntax crate, which covers the lalrpop error
type and lexer error type.

Remove uses of unwrap() from syntax, instead bubble errors up to the
main module, where it can choose what to do. Also changes all unwrap()
calls in main to expect().

Adds the --all flag to travis commands to work with workspace crates.

Move the top of the ast node into the ast submodule, the syntax library
now creates an ast through a function instead of a method of a struct.

Reorganize tests in the following way:

	Unit tests stay the same, integration tests now dispatched from
	'/tests/test.rs', which references silica and verilog sources from
	subfolders.

	'pass-output' contains sources which compile successfully and have
	exact output that is asserted against.

	'fail' contains sources which fail compilation.

	Test utilities for commonly used functions are in '/tests/util/'.