Change verilog module to subcrate
Create a verilog ast module, allowing for easy manipulation of a verilog
ast.

Add display implementations for all nodes of verilog ast for easy
conversion into verilog code.

Add a module that transforms a silica ast into a verilog ast. In this
future this may be better situated into another subcrate, keeping this
one more 'pure', but works for now (especially as there is very little
logic in the silica to verilog converstion, at the moment).